Move use-case tests to the right package
